Move snapshotter to keeper and implement
